Mogudes varies funcions des de `import_fmts` a `ampacsv`
